// apps/motrix/ui/generated/native-downloads.ts
// AUTO-GENERATED by wabou-bindgen. DO NOT EDIT.
export type InspectTorrentRequest = {
	path: string,
};

export type TorrentFilePreview = {
	/**  User-facing BitTorrent file index. These indices are one-based. */
	index: number,
	path: string,
	length: number,
};

export type TorrentPreview = {
	name: string,
	totalLength: number,
	files: TorrentFilePreview[],
};

export interface NativeDownloadsApi {
  inspectTorrent(request: InspectTorrentRequest): Promise<TorrentPreview>;
}